In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,24 +11,24 @@
 
 int main(void)
 {
-	long long int pro1, pro2 = 2, pro3 = 1, i = 1;
+	long int pro1, pro2 = 2, pro3 = 1, i = 1;
 
 	printf("1, ");
 	printf("2, ");
 	while (i <= 32)
 	{
 		pro1 = pro2 + pro3;
-		printf("%llu, ", pro1);
+		printf("%lu, ", pro1);
 		pro3 = pro1 + pro2;
-		printf("%llu, ", pro3);
+		printf("%lu, ", pro3);
 		pro2 = pro3 + pro1;
 		if (i != 32)
 		{
-			printf("%llu, ", pro2);
+			printf("%lu, ", pro2);
 		}
 		else
 		{
-			printf("%llu\n", pro2);
+			printf("%lu\n", pro2);
 			break;
 		}
 		i++;
